invites/requirements.txt
bottle.py
paste.py
os
